% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_resolved_names.R
\name{cache_resolved_names}
\alias{cache_resolved_names}
\title{Cache resolved names}
\usage{
cache_resolved_names(data_vec)
}
\arguments{
\item{data_vec}{A named list containing the resolved names (data.frames).}
}
\description{
Takes a named list with resolved names and caches the results by writing
each unique resolved name to a separate text file.
}
