% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accepted_row.R
\name{get_accepted_row}
\alias{get_accepted_row}
\title{Get the first row with status "ACCEPTED" from a data source}
\usage{
get_accepted_row(data_source, only_accepted = TRUE)
}
\arguments{
\item{data_source}{The data source to filter}

\item{only_accepted}{Logical indicating whether to filter for "ACCEPTED" status}
}
\value{
The first row with status "ACCEPTED" or the first row if "status" column
is not present
}
\description{
This function filters the data source to retrieve the first row
with the status "ACCEPTED".
If the data source does not have a "status" column, it returns the first row.
}
