% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cached_file_names.R
\name{get_cached_file_names}
\alias{get_cached_file_names}
\title{Get the names of cached files in the taxospace directory.}
\usage{
get_cached_file_names(sel_dir)
}
\arguments{
\item{sel_dir}{A string indicating the directory to search for cached files.}
}
\value{
A character vector containing the names of the cached files.
}
\description{
This function retrieves the names of cached files in the taxospace directory.
}
