% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\title{Get classification for a taxonomic name}
\usage{
get_classification(taxa, interactive = TRUE)
}
\arguments{
\item{taxa}{A character vector of taxonomic names}

\item{interactive}{A logical value indicating whether to ask the user for
input when there is no match found in the GBIF database}
}
\value{
A list with the following elements:
\item{sel_name}{The selected taxonomic name}
\item{data_resolve}{A data.frame with information about the resolved name,
including the matched name, data source title, and match score}
\item{db}{The database from which the classification was retrieved
(GBIF or ITIS)}
\item{id}{The most matching ID of the taxonomic name in the database}
\item{classification}{A data.frame with the classification of the taxonomic
name, including the name, rank, and ID}
}
\description{
This function takes a taxonomic name and returns its classification.
It retrieves the classification from the Global Biodiversity Information
Facility (GBIF) and Integrated Taxonomic Information System (ITIS)
databases. If the name is not found in either database,
an empty list is returned.
}
\examples{
get_classification("Homo sapiens")
get_classification(c("Canis lupus", "Felis catus"))
get_classification("Pikachu")

}
