% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_matching_id.R
\name{get_most_matching_id}
\alias{get_most_matching_id}
\title{Get the most matching ID}
\usage{
get_most_matching_id(data_with_nested_resolve, data_taxa_mached_name_id)
}
\arguments{
\item{data_with_nested_resolve}{A data frame containing the nested column \code{data_resolve}.}

\item{data_taxa_mached_name_id}{A data frame containing the column \code{matched_name}.}
}
\value{
A modified data frame with the most matching ID.
}
\description{
This function takes two data frames as input:
\code{data_with_nested_resolve} and \code{data_taxa_mached_name_id}.
It performs a series of operations to find the most matching ID based on
the frequncy of the matched names.
}
