% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_taxa.R
\name{link_taxa}
\alias{link_taxa}
\title{Get common classification for two taxons}
\usage{
link_taxa(taxon_a, taxon_b, ...)
}
\arguments{
\item{taxon_a}{A character with a taxonomic name}

\item{taxon_b}{A character with a taxonomic name}

\item{...}{Additional arguments pased to \link{get_classification}}
}
\value{
A list with the following elements:
\item{taxon_a}{A data frame with the classification of taxon_a}
\item{taxon_b}{A data frame with the classification of taxon_b}
\item{common_classification}{A data frame with the common classification}
\item{highest_common_level}{The highest common classification level}
\item{highest_common_level_name}{The name of the highest common classification level}
\item{highest_common_level_id}{The ID of the highest common classification level}
}
\description{
Get the individuial classification for taxons and then find
the most releated common classification level.
}
\examples{
link_taxa(taxon_a = "Homo sapiens", taxon_b = "Homo denisova")
link_taxa("Canis lupus", "Felis catus")
}
\seealso{
\link{get_classification}
}
