% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cached_classification.R
\name{load_cached_classification}
\alias{load_cached_classification}
\title{Load cached classification}
\usage{
load_cached_classification(data_vector)
}
\arguments{
\item{data_vector}{A vector of file names to load the cached classification names from.}
}
\value{
A data frame containing the loaded data.
}
\description{
This function takes a data vector as input and loads the cached
classification
}
