% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cached_resolved_names.R
\name{load_cached_resolved_names}
\alias{load_cached_resolved_names}
\title{Load cached resolved names}
\usage{
load_cached_resolved_names(data_vector)
}
\arguments{
\item{data_vector}{A vector of file names to load the cached resolved names from.}
}
\value{
A data frame containing the loaded data.
}
\description{
This function takes a data vector as input and loads the cached resolved
names.
}
