% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_resolve.R
\name{select_best_resolve}
\alias{select_best_resolve}
\title{Selects the best resolve}
\usage{
select_best_resolve(data_resolve)
}
\arguments{
\item{data_resolve}{The input data frame containing the resolves.}
}
\value{
A data frame with the selected resolves.
}
\description{
Filters the rows based on the maximum score for for each group.
}
