% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxospace-package.R
\docType{package}
\name{taxospace-package}
\alias{taxospace}
\alias{taxospace-package}
\title{taxospace: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ondrej Mottl \email{ondrej.mottl@gamil.com} (\href{https://orcid.org/0000-0002-9796-5081}{ORCID})

}
\keyword{internal}
