% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\title{Get classification for given vector of taxa}
\usage{
get_classification(
  taxa_vec,
  sel_db_name = c("gnr", "iplant"),
  sel_db_class = c("gbif", "itis"),
  use_cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{taxa_vec}{A character vector of taxa names to classify.}

\item{sel_db_name}{A string indicating which database to use to
resolve names. Default is "gnr".}

\item{sel_db_class}{A string indicating which database to use for
classification. Default is "gbif".}

\item{use_cache}{Logical. If TRUE the intermediate results are cached.
This speeds up the process drastically for large vectors. Default is TRUE.}

\item{verbose}{Logical. If TRUE the additional messages are printed on the console.}
}
\value{
A data.frame with the classification of the taxonomic
name, including the name, rank, and ID
}
\description{
This function takes a vector of taxa names and returns the classification
for each of them using the specified database. The default is to use the
GBIF database, but it can be changed to ITIS with the "sel_db_class"
parameter. If the name is not found in either database,
an empty list is returned.
}
\examples{
get_classification("Homo sapiens")
get_classification(c("Canis lupus", "Felis catus"))
get_classification("Pikachu")
}
\seealso{
\link[taxize:resolve]{taxize::resolve}, \link[taxize:classification]{taxize::classification}
}
