% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\title{Get classification for a taxonomic name}
\usage{
get_classification(
  taxon,
  sel_db_name = c("gnr", "iplant"),
  interactive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{taxon}{A character with a taxonomic name}

\item{sel_db_name}{A string indicating which database to use to
resolve names. Default is "gnr".}

\item{interactive}{A logical value indicating whether to ask the user for
input or automaticaly pick the best match found in the database}

\item{verbose}{Logical. If TRUE the additional messages are printed on
the console}
}
\value{
A list with the following elements:
\item{sel_name}{The selected taxonomic name}
\item{data_resolve}{A data.frame with information about the resolved name,
including the matched name, data source title, and match score}
\item{db}{The database from which the classification was retrieved
(GBIF or ITIS)}
\item{id}{The most matching ID of the taxonomic name in the database}
\item{classification}{A data.frame with the classification of the taxonomic
name, including the name, rank, and ID}
}
\description{
This function takes a taxonomic name and returns its classification.
It retrieves the classification from the Global Biodiversity Information
Facility (GBIF) and Integrated Taxonomic Information System (ITIS)
databases. If the name is not found in either database,
an empty list is returned.
}
\examples{
get_classification("Homo sapiens")
get_classification(c("Canis lupus", "Felis catus"))
get_classification("Pikachu")
}
\seealso{
\link{get_classification_buk}, \link[taxize:resolve]{taxize::resolve}, \link[taxize:classification]{taxize::classification}
}
